Set up event page and components.
EventsPage displays a list of all events via the EventListDisplay component.

Set up CreateEventComponent, which includes a Modal form for the user to fill out.

EventPage displays the information of a specific event via the EventDisplay component.

UpdateEventComponent allows the owner of the event to update the information of the event and even remove other members of the event via a Modal form.

JoinEventComponent available to non-members.

LeaveEventComponent available to members.

guestInit no longer a dependency for the useEffect hook in protected-route