Implement notifications and likes via socket.
Notification event occurs when user receives a post on their own page, a message is sent to the user, or a friend request is sent to the user.

Send friend request was not refactored. Creating a new friend request is still a RESTful process, but there is an adjacent socket process that pings the notification chatroom for the receiving user.

Liking a post and comment has been removed from the PostController and CommentController. It is no longer a route.

Posts and comment likes are handled as a socket event and mongo is updated within this event.