Set up notifications.
New notification properties in model: type and triggeredBy

New notification controller functions to fetch notifications and to fetch notifications and update the read property to true.

Notifications created when user is sent a message, a post is made on the user's page, and the user is sent a friend request.