Implement notifications.
Create NotificationsContext to check for new notifications and update the notifications button UI if there are unread notifications.

Notifications UI in Header component is now updated with sockets and RESTful checks.

Create useNotifications hook to fetch notifications and update the notifications status to read from unread.

NotificationsList component displays notifications on click. Clicking on a notification will take the user to the relevant page.